feature: like manga endpoint     save current user as manga owner when creating a new manga;     add throwNotFound option in the abstract.repository;     fix no default message for UnprocessableEntityException;

+++ express-manga-api/src/services/manga.service.js
@@ -3,6 +3,7 @@ import { findAll, findById, update } from "../database/abstract.repository.js";
 import Chapter from "../models/chapter.model.js";
 import Manga from "../models/manga.model.js";
 import Review from "../models/review.model.js";
+import User from "../models/user.model.js";
 import {
   InternalServerErrorException,
   NotFoundException,
@@ -12,22 +13,66 @@ import { deleteFiles } from "../utils/files.util.js";
 import { getMessage } from "../utils/message.util.js";
 
 async function create(data) {
+  const session = await mongoose.startSession();
   try {
+    session.startTransaction();
+
     const newManga = await Manga.create({
       ...data,
-    });
+    }).session(session);
+
+    const currentUser = await User.findByIdAndUpdate(
+      data.owner,
+      { $push: { mangas: newManga._id } },
+      { session },
+    );
+
+    if (!currentUser) {
+      throw new UnprocessableEntityException();
+    }
+
+    await session.commitTransaction();
+    console.log("Transaction committed successfully");
 
     return newManga;
   }
   catch (err) {
-    await deleteFiles(data.files);
+    console.error("Error:", err);
+
+    await session.abortTransaction();
+
+    if (err instanceof UnprocessableEntityException)
+      throw err;
 
     if (err.code === 11000) {
       throw new UnprocessableEntityException(getMessage("manga.error.twinned"));
     }
+
     throw new InternalServerErrorException(
       "Error while creating manga",
     );
+  }
+  finally {
+    session.endSession();
+  }
+}
+
+async function likeManga(mangaId, data, throwNotFound = true) {
+  try {
+    const document = await Manga.findByIdAndUpdate(
+      mangaId,
+      { $push: { likes: data.userId } },
+    );
+
+    if (!document && throwNotFound) {
+      throw new NotFoundException();
+    }
+
+    return document;
+  }
+  catch (err) {
+    console.error("Error:", err);
+    throw err;
   }
 }
 
@@ -56,18 +101,29 @@ async function deleteById(mangaId, throwNotFound = true) {
     }
     console.log("Manga deleted:", document ? document.title : "Not found");
 
-    // 5. Collect all image files to delete
+    // 5. Delete manga from user list
+    const mangaOwner = await User.findByIdAndUpdate(
+      document.owner,
+      { $pull: { mangas: mangaId } },
+      { session },
+    );
+
+    if (!mangaOwner) {
+      throw new UnprocessableEntityException();
+    }
+
+    // 6. Collect all image files to delete
     const allImages = chapters.flatMap(chapter => chapter.files).concat(document.covers);
 
     console.log({ chapters, document });
 
     console.log("Total images to delete:", allImages.length);
 
-    // 6. Commit transaction first
+    // 7. Commit transaction first
     await session.commitTransaction();
     console.log("Transaction committed successfully");
 
-    // 7. Delete files AFTER successful DB operations
+    // 8. Delete files AFTER successful DB operations
     if (allImages.length > 0) {
       try {
         await deleteFiles(allImages);
@@ -91,9 +147,12 @@ async function deleteById(mangaId, throwNotFound = true) {
 }
 
 export default {
+
   create,
   findById: id => findById(Manga, id),
   findAll: (filter, populate = null) => findAll(Manga, filter, populate),
   update: (filter, data) => update(Manga, filter, data),
   deleteById,
+
+  likeManga,
 };
